+++ astro.config.mjs
@@ -1,10 +1,10 @@
 // @ts-check
-import { defineConfig } from 'astro/config';
-import tailwind from '@astrojs/tailwind';
+import { defineConfig } from "astro/config";
+import tailwind from "@astrojs/tailwind";
 
 // https://astro.build/config
 export default defineConfig({
   integrations: [tailwind()],
-  output: 'static',
-  site: 'https://example.com', // Cambiar por la URL real del sitio
+  output: "static",
+  site: "https://example.com", // Cambiar por la URL real del sitio
 });